Projectile Changes
made changes to projectile and bullet

+++ 2D Tile Game/src/com/winfirst/projectile/Vector2D.java
@@ -2,49 +2,49 @@ package com.winfirst.projectile;
 
 public class Vector2D {
 
-    private double x, y, speed, angle, xVol, yVol;
+    private float x, y, speed, angle, xVol, yVol;
 
-    public Vector2D(double x, double y, double speed) {
+    public Vector2D(float x, float y, float speed) {
         this.x = x;
         this.y = y;
         this.speed = speed;
-        this.angle = Math.atan2(y, x);
-        xVol = (speed) * Math.cos(angle);
-        yVol = (speed) * Math.sin(angle);
+        this.angle = (float)(Math.atan2(y, x));
+        xVol = (float)((speed) * Math.cos(angle));
+        yVol = (float)((speed) * Math.sin(angle));
     }
 
-    public double getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(double x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public double getY() {
+    public float getY() {
         return y;
     }
 
-    public double getXVol(){
+    public float getXVol(){
         return xVol;
     }
 
-    public double getYVol(){
+    public float getYVol(){
         return yVol;
     }
 
-    public void setY(double y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public double getSpeed() {
+    public float getSpeed() {
         return speed;
     }
 
-    public void setSpeed(double speed) {
+    public void setSpeed(float speed) {
         this.speed = speed;
     }
 
-    public double getAngle(){return angle;}
+    public float getAngle(){return angle;}
 
 }
